+++ app/javascript/app/controllers/offer_controller.js
@@ -2,11 +2,10 @@ import {Controller} from 'stimulus'
 import initChoises from "../choises";
 
 export default class extends Controller {
-  static targets = ["parameters", "webpage", "external",
+  static targets = ["parameters", "external",
                     "attributes", "button", "attributeType"];
 
   initialize() {
-    // this.showWebpage();
     this.indexCounter = 0;
     if (this.attributesTarget.firstElementChild) {
       this.attributesTarget.classList.add("active");
@@ -76,14 +75,4 @@ export default class extends Controller {
       current.parentNode.insertBefore(next, current);
     }
   }
-
-  // showWebpage(event){
-  //   const offerType = this.offerTypeTarget.value
-  //   const external = this.externalTarget.checked
-  //   if (offerType !== "order_required" || external ){
-  //     this.webpageTarget.classList.remove("hidden-fields");
-  //   } else {
-  //     this.webpageTarget.classList.add("hidden-fields");
-  //   }
-  // }
 }