+++ app/policies/backoffice/offer_policy.rb
@@ -44,7 +44,7 @@ class Backoffice::OfferPolicy < ApplicationPolicy
   end
 
   def permitted_attributes
-    [:id, :name, :description, :webpage, :order_type, :order_url, :internal,
+    [:id, :name, :description, :order_type, :order_url, :internal,
      :primary_oms_id, oms_params: {},
      parameters_attributes: [:type, :name, :hint, :min, :max,
                              :unit, :value_type, :start_price, :step_price, :currency,