+++ app/services/service/update.rb
@@ -10,14 +10,12 @@ class Service::Update
     if @service.offers.size == 1
       Offer::Update.new(@service.offers.first, { order_type: @params[:order_type] || @service.order_type,
                                    order_url: @params[:order_url] || @service.order_url,
-                                   webpage: @params[:webpage_url] || @service.webpage_url,
                                    status: "published" }).call
     elsif @service.offers.size == 0
       Offer::Create.new(Offer.new(name: "Offer",
                                   description: "#{@params[:name] || @service.name} Offer",
                                   order_type: @params[:order_type] || @service.order_type,
                                   order_url: @params[:order_url] || @service.order_url,
-                                  webpage: @params[:webpage_url] || @service.webpage_url,
                                   status: "published",
                                   service: @service)).call
     end