+++ CHANGELOG.md
@@ -14,11 +14,13 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Resource provider can navigate to the ordering parameters management from the Resource Presentation Page (@kmarszalek, @jarekzet)
 
 ### Changed
+- For an Offer, an empty order URL doesn't imply internal ordering (@jswk)
 - More consistent form behaviour for ordering configuration (@jswk)
 
 ### Deprecated
 
 ### Removed
+- Offer webpage field (@jswk)
 
 ### Fixed